Ensure that A1 & A2 are present, if not can find it with SNP and other allele

More care needs to be taken if one of A1/A2 is present, before imputing the other allele flipping needs to be checked

check_no_allele(
  sumstats_dt,
  path,
  ref_genome,
  rsids,
  imputation_ind,
  allele_flip_check,
  log_folder_ind,
  check_save_out,
  tabix_index,
  nThread,
  log_files,
  bi_allelic_filter
)

Arguments

path
Filepath for the summary statistics file to be formatted. A dataframe or datatable of the summary statistics file can also be passed directly to MungeSumstats using the path parameter.

ref_genome
name of the reference genome used for the GWAS ("GRCh37" or "GRCh38"). Argument is case-insensitive. Default is NULL which infers the reference genome from the data.

imputation_ind
Binary Should a column be added for each imputation step to show what SNPs have imputed values for differing fields. This includes a field denoting SNP allele flipping (flipped). On the flipped value, this denoted whether the alelles where switched based on MungeSumstats initial choice of A1, A2 from the input column headers and thus may not align with what the creator intended.Note these columns will be in the formatted summary statistics returned. Default is FALSE.

allele_flip_check
Binary Should the allele columns be checked against reference genome to infer if flipping is necessary. Default is TRUE.

log_folder_ind
Binary Should log files be stored containing all filtered out SNPs (separate file per filter). The data is outputted in the same format specified for the resulting sumstats file. The only exception to this rule is if output is vcf, then log file saved as .tsv.gz. Default is FALSE.

tabix_index
Index the formatted summary statistics with tabix for fast querying.

nThread
Number of threads to use for parallel processes.

log_files
list of log file locations

bi_allelic_filter
Binary Should non-biallelic SNPs be removed. Default is TRUE.

Value

A list containing two data tables:

sumstats_dt: the modified summary statistics data table object

rsids: snpsById, filtered to SNPs of interest if loaded already. Or else NULL

allele_flip_check: does the dataset require allele flip check

log_files: log file list

bi_allelic_filter: should multi-allelic SNPs be filtered out