Add user supplied sample size

Add user supplied sample size

compute_sample_size_n(sumstats_dt, method, force_new = FALSE)

Arguments

sumstats_dt

Summary statistics data.table.

method

Method for computing (effective) sample size.

"ldsc" :
\(Neff = (N_CAS+N_CON) * (N_CAS/(N_CAS+N_CON)) / mean((N_CAS/(N_CAS+N_CON))[(N_CAS+N_CON)==max(N_CAS+N_CON)]))\)
bulik/ldsc GitHub Issue bulik/ldsc GitHub code

"giant" :
\(Neff = 2 / (1/N_CAS + 1/N_CON)\)
Winkler et al. 2014, Nature

"metal" :
\(Neff = 4 / (1/N_CAS + 1/N_CON)\)
Willer et al. 2010, Bioinformatics

"sum" :
\(N = N_CAS + N_CON\)
Simple summation of cases and controls that does not account for class imbalance.

"\<integer\>" :
N = \<integer\>
If method is a positive integer, it will be used as N for every row.

force_new

If "Neff" (or "N") already exists in sumstats_dt, replace it with the recomputed version.

Value

No return