Search Open GWAS for datasets matching criteria

For each argument, searches for any datasets matching a case-insensitive substring search in the respective metadata column. Users can supply a single character string or a list/vector of character strings.

find_sumstats(
  ids = NULL,
  traits = NULL,
  years = NULL,
  consortia = NULL,
  authors = NULL,
  populations = NULL,
  categories = NULL,
  subcategories = NULL,
  builds = NULL,
  pmids = NULL,
  min_sample_size = NULL,
  min_ncase = NULL,
  min_ncontrol = NULL,
  min_nsnp = NULL,
  include_NAs = FALSE,
  access_token = check_access_token()
)

Arguments

ids
List of Open GWAS study IDs (e.g. c("prot-a-664", "ieu-b-4760")).

traits
List of traits (e.g. c("parkinson", "Alzheimer")).

years
List of years (e.g. seq(2015,2021) or c(2010, 2012, 2021)).

consortia
List of consortia (e.g. c("MRC-IEU","Neale Lab").

authors
List of authors (e.g. c("Elsworth","Kunkle","Neale")).

populations
List of populations (e.g. c("European","Asian")).

categories
List of categories (e.g. c("Binary","Continuous","Disease","Risk factor"))).

subcategories
List of categories (e.g. c("neurological","Immune","cardio"))).

builds
List of genome builds (e.g. c("hg19","grch37")).

pmids
List of PubMed ID (exact matches only) (e.g. c(29875488, 30305740, 28240269)).

min_sample_size
Minimum total number of study participants (e.g. 5000).

min_ncase
Minimum number of case participants (e.g. 1000).

min_ncontrol
Minimum number of control participants (e.g. 1000).

min_nsnp
Minimum number of SNPs (e.g. 200000).

include_NAs
Include datasets with missing metadata for size criteria (i.e. min_sample_size, min_ncase, or min_ncontrol).

access_token
Google OAuth2 access token. Used to authenticate level of access to data

Value

(Filtered) GWAS metadata table.

Details

By default, returns metadata for all studies currently in Open GWAS database.

Examples

### By ID
metagwas <- MungeSumstats::find_sumstats(ids=c("ieu-b-4760","prot-a-1725","prot-a-664" ))

#> Collecting metadata from Open GWAS.

#> Filtering metadata by substring criteria.

#> Found 3 GWAS datasets matching search criteria across:
#>    - 3 trait(s)
#>    - 1 population(s)
#>    - 2 category(ies)
#>    - 2 subcategory(ies)
#>    - 2 publication(s)
#>    - 2 consortia(ium)
#>    - 1 genome build(s)

### By ID amd sample size
metagwas <- MungeSumstats::find_sumstats(ids=c("ieu-b-4760","prot-a-1725","prot-a-664"),
                                         min_sample_size=5000)

#> Collecting metadata from Open GWAS.

#> Filtering metadata by substring criteria.

#> Filtering metadata by sample/case/control/SNP size criteria.

#> Excluding sample/case/control size with NAs.

#> Found 1 GWAS datasets matching search criteria across:
#>    - 1 trait(s)
#>    - 1 population(s)
#>    - 1 category(ies)
#>    - 1 subcategory(ies)
#>    - 1 publication(s)
#>    - 1 consortia(ium)
#>    - 1 genome build(s)

### By criteria
metagwas <- MungeSumstats::find_sumstats(traits=c("alzheimer","parkinson"),
                                         years = seq(2015,2021)) 

#> Collecting metadata from Open GWAS.

#> Filtering metadata by substring criteria.

#> Found 29 GWAS datasets matching search criteria across:
#>    - 21 trait(s)
#>    - 1 population(s)
#>    - 2 category(ies)
#>    - 2 subcategory(ies)
#>    - 3 publication(s)
#>    - 5 consortia(ium)
#>    - 1 genome build(s)