Infer genome builds

Infers the genome build of summary statistics files (GRCh37 or GRCh38) from the data. Uses SNP (RSID) & CHR & BP to get genome build.

get_genome_builds(
  sumstats_list,
  header_only = TRUE,
  sampled_snps = 10000,
  names_from_paths = FALSE,
  nThread = 1
)

Arguments

sumstats_list
A named list of paths to summary statistics, or a named list of data.table objects.

header_only
Instead of reading in the entire sumstats file, only read in the first N rows where N=sampled_snps. This should help speed up cases where you have to read in sumstats from disk each time.

sampled_snps
Downsample the number of SNPs used when inferring genome build to save time.

names_from_paths
Infer the name of each item in sumstats_list from its respective file path. Only works if sumstats_list is a list of paths.

nThread
Number of threads to use for parallel processes.

Value

ref_genome the genome build of the data

Details

Iterative version of get_genome_build.

Examples

# Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata", "eduAttainOkbay.txt",
    package = "MungeSumstats"
)
sumstats_list <- list(ss1 = eduAttainOkbayPth, ss2 = eduAttainOkbayPth)

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks
is_32bit_windows <-
    .Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
    ref_genomes <- get_genome_builds(sumstats_list = sumstats_list)
}

#> Inferring genome build of 2 sumstats file(s).

#> Inferring genome build.

#> Reading in only the first 10000 rows of sumstats.

#> Reading header.

#> Tabular format detected.

#> Importing tabular file: /private/var/folders/hd/jm8lzp7s4dl_wlkykzhz66x80000gn/T/RtmpybTkfG/temp_libpath112a4254b8072/MungeSumstats/extdata/eduAttainOkbay.txt

#> Standardising column headers.

#> First line of summary statistics file: 

#> MarkerName	CHR	POS	A1	A2	EAF	Beta	SE	Pval

#> Loading reference genome data.

#> Loading reference genome data.

#> Inferred genome build: GRCH37

#> Inferring genome build.

#> Reading in only the first 10000 rows of sumstats.

#> Reading header.

#> Tabular format detected.

#> Importing tabular file: /private/var/folders/hd/jm8lzp7s4dl_wlkykzhz66x80000gn/T/RtmpybTkfG/temp_libpath112a4254b8072/MungeSumstats/extdata/eduAttainOkbay.txt

#> Standardising column headers.

#> First line of summary statistics file: 

#> MarkerName	CHR	POS	A1	A2	EAF	Beta	SE	Pval

#> Loading reference genome data.

#> Loading reference genome data.

#> Inferred genome build: GRCH37

#> Time difference of 1.251583 mins

#> GRCH37: 2 file(s)