Genome build liftover

Transfer genomic coordinates from one genome build to another.

liftover(
  sumstats_dt,
  convert_ref_genome,
  ref_genome,
  imputation_ind,
  verbose = TRUE
)

Source

liftOver

UCSC chain files

Arguments

sumstats_dt

data table obj of the summary statistics file for the GWAS

convert_ref_genome

name of the reference genome to convert to ("GRCh37" or "GRCh38"). This will only occur if the current genome build does not match. Default is not to convert the genome build (NULL).

ref_genome

name of the reference genome used for the GWAS ("GRCh37" or "GRCh38"). Argument is case-insensitive. Default is NULL which infers the reference genome from the data.

imputation_ind

Binary Should a column be added for each imputation step to show what SNPs have imputed values for differing fields. This includes a field denoting SNP allele flipping (flipped). On the flipped value, this denoted whether the alelles where switched based on MungeSumstats initial choice of A1, A2 from the input column headers and thus may not align with what the creator intended.Note these columns will be in the formatted summary statistics returned. Default is FALSE.

verbose

Print messages.

Value

list("sumstats_dt"=sumstats_dt)