Check that summary statistics from GWAS are in a homogeneous format

Check that summary statistics from GWAS are in a homogeneous format

format_sumstats(
  path,
  ref_genome = NULL,
  convert_ref_genome = NULL,
  convert_small_p = TRUE,
  compute_z = FALSE,
  force_new_z = FALSE,
  compute_n = 0L,
  convert_n_int = TRUE,
  analysis_trait = NULL,
  INFO_filter = 0.9,
  FRQ_filter = 0,
  pos_se = TRUE,
  effect_columns_nonzero = FALSE,
  N_std = 5,
  N_dropNA = TRUE,
  rmv_chr = c("X", "Y", "MT"),
  rmv_chrPrefix = TRUE,
  on_ref_genome = TRUE,
  strand_ambig_filter = FALSE,
  allele_flip_check = TRUE,
  allele_flip_drop = TRUE,
  allele_flip_z = TRUE,
  allele_flip_frq = TRUE,
  bi_allelic_filter = TRUE,
  snp_ids_are_rs_ids = TRUE,
  remove_multi_rs_snp = FALSE,
  sort_coordinates = TRUE,
  nThread = 1,
  save_path = tempfile(fileext = ".tsv.gz"),
  write_vcf = FALSE,
  tabix_index = FALSE,
  return_data = FALSE,
  return_format = "data.table",
  ldsc_format = FALSE,
  log_folder_ind = FALSE,
  log_mungesumstats_msgs = FALSE,
  log_folder = tempdir(),
  imputation_ind = FALSE,
  force_new = FALSE,
  mapping_file = sumstatsColHeaders
)

Arguments

path
Filepath for the summary statistics file to be formatted. A dataframe or datatable of the summary statistics file can also be passed directly to MungeSumstats using the path parameter.

ref_genome
name of the reference genome used for the GWAS ("GRCh37" or "GRCh38"). Argument is case-insensitive. Default is NULL which infers the reference genome from the data.

convert_ref_genome
name of the reference genome to convert to ("GRCh37" or "GRCh38"). This will only occur if the current genome build does not match. Default is not to convert the genome build (NULL).

convert_small_p
Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.

compute_z
Whether to compute Z-score column from P. Default is FALSE. Note that imputing the Z-score for every SNP will not correct be perfectly correct and may result in a loss of power. This should only be done as a last resort.

force_new_z
When a "Z" column already exists, it will be used by default. To override and compute a new Z-score column from P set force_new_z=TRUE.

compute_n
Whether to impute N. Default of 0 won't impute, any other integer will be imputed as the N (sample size) for every SNP in the dataset. Note that imputing the sample size for every SNP is not correct and should only be done as a last resort. N can also be inputted with "ldsc", "sum", "giant" or "metal" by passing one of these for this field or a vector of multiple. Sum and an integer value creates an N column in the output whereas giant, metal or ldsc create an Neff or effective sample size. If multiples are passed, the formula used to derive it will be indicated.

convert_n_int
Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.

analysis_trait
If multiple traits were studied, name of the trait for analysis from the GWAS. Default is NULL.

INFO_filter
numeric The minimum value permissible of the imputation information score (if present in sumstats file). Default 0.9.

FRQ_filter
numeric The minimum value permissible of the frequency(FRQ) of the SNP (i.e. Allele Frequency (AF)) (if present in sumstats file). By default no filtering is done, i.e. value of 0.

pos_se
Binary Should the standard Error (SE) column be checked to ensure it is greater than 0? Those that are, are removed (if present in sumstats file). Default TRUE.

effect_columns_nonzero
Binary should the effect columns in the data BETA,OR (odds ratio),LOG_ODDS,SIGNED_SUMSTAT be checked to ensure no SNP=0. Those that do are removed(if present in sumstats file). Default FALSE.

N_std
numeric The number of standard deviations above the mean a SNP's N is needed to be removed. Default is 5.

N_dropNA
Drop rows where N is missing.Default is TRUE.

rmv_chr
vector or character The chromosomes on which the SNPs should be removed. Use NULL if no filtering necessary. Default is X, Y and mitochondrial.

rmv_chrPrefix
Remove "chr" or "CHR" from chromosome names. Default is TRUE.

on_ref_genome
Binary Should a check take place that all SNPs are on the reference genome by SNP ID. Default is TRUE.

strand_ambig_filter
Binary Should SNPs with strand-ambiguous alleles be removed. Default is FALSE.

allele_flip_check
Binary Should the allele columns be checked against reference genome to infer if flipping is necessary. Default is TRUE.

allele_flip_drop
Binary Should the SNPs for which neither their A1 or A2 base pair values match a reference genome be dropped. Default is TRUE.

allele_flip_z
Binary should the Z-score be flipped along with effect and FRQ columns like Beta? It is assumed to be calculated off the effect size not the P-value and so will be flipped i.e. default TRUE.

allele_flip_frq
Binary should the frequency (FRQ) column be flipped along with effect and z-score columns like Beta? Default TRUE.

bi_allelic_filter
Binary Should non-biallelic SNPs be removed. Default is TRUE.

snp_ids_are_rs_ids
Binary Should the supplied SNP ID's be assumed to be RS IDs. If not, imputation using the SNP ID for other columns like base-pair position or chromosome will not be possible. If set to FALSE, the SNP RS ID will be imputed from the reference genome if possible. Default is TRUE.

remove_multi_rs_snp
Binary Sometimes summary statistics can have multiple RS IDs on one row (i.e. related to one SNP), for example "rs5772025_rs397784053". This can cause an error so by default, the first RS ID will be kept and the rest removed e.g."rs5772025". If you want to just remove these SNPs entirely, set it to TRUE. Default is FALSE.

sort_coordinates
Whether to sort by coordinates.

nThread
Number of threads to use for parallel processes.

save_path
File path to save formatted data. Defaults to tempfile(fileext=".tsv.gz").

write_vcf
Whether to write as VCF (TRUE) or tabular file (FALSE).

tabix_index
Index the formatted summary statistics with tabix for fast querying.

return_data
Return data.table, GRanges or VRanges directly to user. Otherwise, return the path to the save data. Default is FALSE.

return_format
If return_data is TRUE. Object type to be returned ("data.table","vranges","granges").

ldsc_format
Binary Ensure that output format meets all requirements to be fed directly into LDSC without the need for additional munging. Default is FALSE

log_folder_ind
Binary Should log files be stored containing all filtered out SNPs (separate file per filter). The data is outputted in the same format specified for the resulting sumstats file. The only exception to this rule is if output is vcf, then log file saved as .tsv.gz. Default is FALSE.

log_mungesumstats_msgs
Binary Should a log be stored containing all messages and errors printed by MungeSumstats in a run. Default is FALSE

log_folder
Filepath to the directory for the log files and the log of MungeSumstats messages to be stored. Default is a temporary directory.

imputation_ind
Binary Should a column be added for each imputation step to show what SNPs have imputed values for differing fields. This includes a field denoting SNP allele flipping (flipped). On the flipped value, this denoted whether the alelles where switched based on MungeSumstats initial choice of A1, A2 from the input column headers and thus may not align with what the creator intended.Note these columns will be in the formatted summary statistics returned. Default is FALSE.

force_new
If a formatted file of the same names as save_path exists, formatting will be skipped and this file will be imported instead (default). Set force_new=TRUE to override this.

mapping_file
MungeSumstats has a pre-defined column-name mapping file which should cover the most common column headers and their interpretations. However, if a column header that is in youf file is missing of the mapping we give is incorrect you can supply your own mapping file. Must be a 2 column dataframe with column names "Uncorrected" and "Corrected". See data(sumstatsColHeaders) for default mapping and necessary format.

Value

The address for the modified sumstats file or the actual data dependent on user choice. Also, if log files wanted by the user, the return in both above instances are a list.

Examples

# Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata", "eduAttainOkbay.txt",
    package = "MungeSumstats"
)

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks

is_32bit_windows <-
    .Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
    reformatted <- format_sumstats(
        path = eduAttainOkbayPth,
        ref_genome = "GRCh37"
    )
} else {
    reformatted <- format_sumstats(
        path = eduAttainOkbayPth,
        ref_genome = "GRCh37",
        on_ref_genome = FALSE,
        strand_ambig_filter = FALSE,
        bi_allelic_filter = FALSE,
        allele_flip_check = FALSE
    )
}

#> 
#> 
#> ******::NOTE::******
#> - Formatted results will be saved to `tempdir()` by default.
#> - This means all formatted summary stats will be deletedupon ending the R session.
#> - To keep formatted summary stats, change `save_path` ( e.g. `save_path=file.path('./formatted',basename(path))` ),  or make sure to copy files elsewhere after processing ( e.g. `file.copy(save_path, './formatted/' )`.
#> ********************

#> Formatted summary statistics will be saved to ==> /var/folders/hd/jm8lzp7s4dl_wlkykzhz66x80000gn/T//RtmpdxFyWB/file17f0a144ca1c.tsv.gz

#> Reading header.

#> Tabular format detected.

#> Importing tabular file: /private/var/folders/hd/jm8lzp7s4dl_wlkykzhz66x80000gn/T/Rtmp8MQyb7/temp_libpath5a45ead8323/MungeSumstats/extdata/eduAttainOkbay.txt

#> Standardising column headers.

#> First line of summary statistics file: 

#> MarkerName	CHR	POS	A1	A2	EAF	Beta	SE	Pval

#> Summary statistics report:
#>    - 93 rows
#>    - 93 unique variants
#>    - 70 genome-wide significant variants (P<5e-8)
#>    - 20 chromosomes

#> Checking for multi-GWAS.

#> Checking for multiple RSIDs on one row.

#> Checking SNP RSIDs.

#> Checking for merged allele column.

#> Ensuring all SNPs are on the reference genome.

#> Loading reference genome data.

#> Checking for correct direction of A1 (reference) and A2 (alternative allele).

#> There are 46 SNPs where A1 doesn't match the reference genome.
#> These will be flipped with their effect columns.

#> Checking for missing data.

#> Checking for duplicate columns.

#> Checking for duplicate SNPs from SNP ID.

#> Checking for SNPs with duplicated base-pair positions

#> Filtering SNPs, ensuring SE>0.

#> Ensuring all SNPs have N<5 std dev above mean.

#> Removing 'chr' prefix from CHR.

#> Making X/Y CHR uppercase.

#> Checking for bi-allelic SNPs.

#> Warning: When method is an integer, must be >0.

#> Sorting coordinates.

#> Writing in tabular format ==> /var/folders/hd/jm8lzp7s4dl_wlkykzhz66x80000gn/T//RtmpdxFyWB/file17f0a144ca1c.tsv.gz

#> Summary statistics report:
#>    - 93 rows (100% of original 93 rows)
#>    - 93 unique variants
#>    - 70 genome-wide significant variants (P<5e-8)
#>    - 20 chromosomes

#> Successfully finished preparing sumstats file, preview:

#>           SNP CHR       BP A1 A2     FRQ   BETA    SE         P
#> 1:   rs301800   1  8490603  T  C 0.17910  0.019 0.003 1.794e-08
#> 2: rs11210860   1 43982527  G  A 0.63060 -0.017 0.003 2.359e-10
#> 3: rs34305371   1 72733610  G  A 0.91231 -0.035 0.005 3.762e-14
#> 4:  rs2568955   1 72762169  T  C 0.23690 -0.017 0.003 1.797e-08
#> 5:  rs1008078   1 91189731  C  T 0.62690  0.016 0.003 6.005e-10

#> Returning path to saved data.

# returned location has the updated summary statistics file