Read in VCF file

Read in a VCF file as a VCF or a data.table. Can optionally save the VCF/data.table as well.

read_vcf(
  path,
  write_vcf = FALSE,
  save_path = NULL,
  tabix_index = FALSE,
  samples = 1,
  which = NULL,
  use_params = TRUE,
  sampled_rows = 1e+07,
  nThread = 1,
  verbose = TRUE
)

Source

#### Benchmarking #### library(VCFWrenchR) library(VariantAnnotation)path <- "https://gwas.mrcieu.ac.uk/files/ubm-a-2929/ubm-a-2929.vcf.gz" vcf <- VariantAnnotation::readVcf(file = path) N <- 1e5 vcf_sub <- vcf[1:N,]res <- microbenchmark::microbenchmark( "vcf2df"={dat1 <- MungeSumstats:::vcf2df(vcf = vcf_sub)}, "VCFWrenchR"= {dat2 <- as.data.frame(x = vcf_sub)}, "VRanges"={dat3 <- data.table::as.data.table( methods::as(vcf_sub, "VRanges"))}, times=1 )

Discussion on VariantAnnotation GitHub

Arguments

path

Filepath for the summary statistics file to be formatted. A dataframe or datatable of the summary statistics file can also be passed directly to MungeSumstats using the path parameter.

write_vcf

Whether to write as VCF (TRUE) or tabular file (FALSE).

save_path

File path to save formatted data. Defaults to tempfile(fileext=".tsv.gz").

tabix_index

Index the formatted summary statistics with tabix for fast querying.

samples

Which samples to use:

1 : Only the first sample will be used (DEFAULT).

NULL : All samples will be used.

c("<sample_id1>","<sample_id2>",...) : Only user-selected samples will be used (case-insensitive).

which

A GRanges describing the sequences and ranges to be queried. Variants whose POS lies in the interval(s) [start, end] are returned. If which is not specified all ranges are returned.

use_params

When TRUE (default), increases the speed of reading in the VCF by omitting columns that are empty based on the head of the VCF (NAs only). NOTE that that this requires the VCF to be sorted, bgzip-compressed, tabix-indexed, which read_vcf will attempt to do.

sampled_rows

First N rows to sample. Set NULL to use full sumstats_file. when determining whether cols are empty.

nThread

Number of threads to use for parallel processes.

verbose

Print messages.

Value

The VCF file in data.table format.

Examples

#### Local file ####
path <- system.file("extdata","ALSvcf.vcf", package="MungeSumstats")
sumstats_dt <- read_vcf(path = path)
#> Compressing and tabix-indexing VCF file.
#> Finding empty VCF columns based on first 1e+07 rows.
#> Converting VCF to data.table.
#> Checking for empty columns.
#> Removing 2 empty columns.
#> Time difference of 0.2 secs
#> 6 empty column(s) detected.
#> 1 sample detected: EBI-a-GCST005647
#> Constructing ScanVcfParam object.
#> Reading VCF file.
#> Time difference of 1.3 secs
#> Converting VCF to data.table.
#> Checking for empty columns.
#> Removing 2 empty columns.
#> Time difference of 0.2 secs
#> Dropping 2 duplicate columns.
#> Checking for empty columns.
#> Unlisting 4 columns.
#> Renaming ID as SNP.
#> VCF file has -log10 P-values, these will be  converted to unadjusted p-values in the 'P' column.
#> No INFO (SI) column detected.

#### Remote file ####
path <- "https://gwas.mrcieu.ac.uk/files/ieu-a-298/ieu-a-298.vcf.gz" 
sumstats_dt2 <- read_vcf(path = path)
#> File inferred to be remote.
#> Finding empty VCF columns based on first 1e+07 rows.
#> Converting VCF to data.table.
#> Checking for empty columns.
#> Removing 2 empty columns.
#> Time difference of 0.1 secs
#> 5 empty column(s) detected.
#> 1 sample detected: ieu-a-298
#> Constructing ScanVcfParam object.
#> Reading VCF file.
#> Time difference of 7.9 secs
#> Converting VCF to data.table.
#> Checking for empty columns.
#> Removing 2 empty columns.
#> Time difference of 0.4 secs
#> Dropping 1 duplicate columns.
#> Checking for empty columns.
#> Removing 1 empty columns.
#> Unlisting 4 columns.
#> Renaming ID as SNP.
#> VCF file has -log10 P-values, these will be  converted to unadjusted p-values in the 'P' column.
#> No INFO (SI) column detected.