Ensure that the input parameters are logical

Ensure that the input parameters are logical

validate_parameters(
  path,
  ref_genome,
  convert_small_p,
  compute_z,
  compute_n,
  convert_n_int,
  analysis_trait,
  INFO_filter,
  pos_se,
  effect_columns_nonzero,
  N_std,
  N_dropNA,
  rmv_chr,
  on_ref_genome,
  strand_ambig_filter,
  allele_flip_check,
  allele_flip_drop,
  allele_flip_z,
  bi_allelic_filter,
  snp_ids_are_rs_ids,
  write_vcf,
  return_format,
  ldsc_format,
  imputation_ind,
  log_folder_ind,
  log_mungesumstats_msgs
)

Arguments

path
Filepath for the summary statistics file to be formatted. Default is a temporary file.

ref_genome
name of the reference genome used for the GWAS ("GRCh37" or "GRCh38"). Argument is case-insensitive. Default is NULL which infers the reference genome from the data.

convert_small_p
Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.

compute_z
Whether to compute Z-score column from P. Default is FALSE. Note that imputing the Z-score for every SNP will not correct be perfectly correct and may result in a loss of power. This should only be done as a last resort.

compute_n
Whether to impute N. Default of 0 won't impute, any other integer will be imputed as the N (sample size) for every SNP in the dataset. Note that imputing the sample size for every SNP is not correct and should only be done as a last resort.

convert_n_int
Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.

analysis_trait
If multiple traits were studied, name of the trait for analysis from the GWAS. Default is NULL.

INFO_filter
numeric The minimum value permissible of the imputation information score (if present in sumstats file). Default 0.9.

pos_se
Binary Should the standard Error (SE) column be checked to ensure it is greater than 0? Those that are, are removed (if present in sumstats file). Default TRUE.

effect_columns_nonzero
Binary should the effect columns in the data BETA,OR (odds ratio),LOG_ODDS,SIGNED_SUMSTAT be checked to ensure no SNP=0. Those that do are removed(if present in sumstats file). Default TRUE.

N_std
numeric The number of standard deviations above the mean a SNP's N is needed to be removed. Default is 5.

N_dropNA
Drop rows where N is missing.Default is TRUE.

rmv_chr
vector or character The chromosomes on which the SNPs should be removed. Use NULL if no filtering necessary. Default is X, Y and mitochondrial.

on_ref_genome
Binary Should a check take place that all SNPs are on the reference genome by SNP ID. Default is TRUE.

strand_ambig_filter
Binary Should SNPs with strand-ambiguous alleles be removed. Default is FALSE.

allele_flip_check
Binary Should the allele columns be checked against reference genome to infer if flipping is necessary. Default is TRUE.

allele_flip_drop
Binary Should the SNPs for which neither their A1 or A2 base pair values match a reference genome be dropped. Default is TRUE.

allele_flip_z
Binary should the Z-score be flipped along with effect columns like Beta? It is assumed to be calculated off the effect size not the P-value and so will be flipped i.e. default TRUE.

bi_allelic_filter
Binary Should non-biallelic SNPs be removed. Default is TRUE.

snp_ids_are_rs_ids
Binary Should the supplied SNP ID's be assumed to be RS IDs. If not, imputation using the SNP ID for other columns like base-pair position or chromosome will not be possible. If set to FALSE, the SNP RS ID will be imputed from the reference genome if possible. Default is TRUE.

write_vcf
Whether to write as VCF (TRUE) or tabular file (FALSE).

return_format
If return_data is TRUE. Object type to be returned ("data.table","vranges","granges").

ldsc_format
Binary Ensure that output format meets all requirements to be fed directly into LDSC without the need for additional munging. Default is FALSE

imputation_ind
Binary Should a column be added for each imputation step to show what SNPs have imputed values for differing fields. This includes a field denoting SNP allele flipping (flipped). On the flipped value, this denoted whether the alelles where switched based on MungeSumstats initial choice of A1, A2 from the input column headers and thus may not align with what the creator intended.Note these columns will be in the formatted summary statistics returned. Default is FALSE.

log_folder_ind
Binary Should log files be stored containing all filtered out SNPs (separate file per filter). The data is outputted in the same format specified for the resulting sumstats file. The only exception to this rule is if output is vcf, then log file saved as .tsv.gz. Default is FALSE.

log_mungesumstats_msgs
Binary Should a log be stored containing all messages and errors printed by MungeSumstats in a run. Default is FALSE

Value

No return